refactor(dpp): apply byte array encoding-stability check in place (match #3865)
Per maintainer decision, drop the version-gating and add the byte-array
encoding-stability check directly to validate_update_v0, consistent with the
sibling chain-halt fix #3865 which added its rejection rule to
validate_structure_v0 in place. The check runs before validate_schema so it
cannot be bypassed by a JSON-schema-compatible widening.

- revert validate_update_v1, the dispatcher arm, and the
  DPP_VALIDATION_VERSIONS_V3 feature-version bump
- check + tests now live in validate_update/v0 (tests still exercise the public
  validate_update dispatcher); drops the pre-activation gating test
- e2e error assertion keeps assert_matches! on the decode-failure variant
